refactor: modularize OAuth handling and improve age calculation
- Created a new OAuthCallbackHandler component to manage Google user profile completion.
- Extracted age calculation logic into a utility function for reuse across components.
- Updated SignUpForm to utilize the new age calculation utility, simplifying the code.
- Cleaned up the TODO list by removing completed tasks and refining remaining items.
