+++ TODO.md
@@ -1,15 +1,3 @@
-- Make a dashboard for each league.
-  - Has to display: - Leaderboard button - DONE Quinipolos to answer - DONE Quinipolos to Correct (if the user is the moderator of this specific league)
-  <!-- - Make screen to show quinipolo correction results. Has to generate a message and must have options to share it automatically. -->
-- Make the teams schema better. It shouldn't be all in the same array. Each should be an entry, with information such as SPORT, LEAGUE (optional) and SEX.
-- Make stronger login with JWT.
-<!-- - Fix handleCloseFeedback -->
-- Chequear el QR code a ver si va.
-- Fix pete al hacer clic en una opción seleccionada en la corrección
-  - Comprobar comportamiento en el formulario de respuesta.
-- Add date to Quinipolo corrected message.
-- Stop using ANTd and use shadcn instead for date handling.
-
 TO DO:
 
 - check the acceptance or declining of petitions for leagues.
