+++ src/utils/calculateAge.ts
@@ -1,0 +1,13 @@
+export function calculateAge(input: string | Date): number {
+  const birthDate = typeof input === "string" ? new Date(input) : input;
+  const today = new Date();
+  let age = today.getFullYear() - birthDate.getFullYear();
+  const monthDifference = today.getMonth() - birthDate.getMonth();
+  if (
+    monthDifference < 0 ||
+    (monthDifference === 0 && today.getDate() < birthDate.getDate())
+  ) {
+    age -= 1;
+  }
+  return age;
+}
